Refactor api paths

+++ rest-api-server/src/main/java/br/com/tiagoamp/timetracker/service/UserService.java
@@ -3,11 +3,10 @@ package br.com.tiagoamp.timetracker.service;
 import br.com.tiagoamp.timetracker.error.ResourceAlreadyRegisteredException;
 import br.com.tiagoamp.timetracker.error.ResourceNotFoundException;
 import br.com.tiagoamp.timetracker.error.TimeTrackerOperationException;
-import br.com.tiagoamp.timetracker.mapper.CategoryMapper;
 import br.com.tiagoamp.timetracker.mapper.UserMapper;
-import br.com.tiagoamp.timetracker.model.Category;
 import br.com.tiagoamp.timetracker.model.User;
-import br.com.tiagoamp.timetracker.repository.*;
+import br.com.tiagoamp.timetracker.repository.UserEntity;
+import br.com.tiagoamp.timetracker.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,19 +18,15 @@ import static java.util.stream.Collectors.toList;
 public class UserService {
 
     private UserRepository userRepo;
-    private CategoryRepository categoryRepo;
-    private TimeEntryRepository timeEntryRepo;
+    private CategoryService categoryService;
     private UserMapper userMapper;
-    private CategoryMapper categoryMapper;
 
 
     @Autowired
-    public UserService(UserRepository userRepo, CategoryRepository categoryRepo, TimeEntryRepository timeEntryRepo, UserMapper userMapper, CategoryMapper categoryMapper) {
+    public UserService(UserRepository userRepo, CategoryService categoryService, UserMapper userMapper) {
         this.userRepo = userRepo;
-        this.categoryRepo = categoryRepo;
-        this.timeEntryRepo = timeEntryRepo;
+        this.categoryService = categoryService;
         this.userMapper = userMapper;
-        this.categoryMapper = categoryMapper;
     }
 
 
@@ -42,16 +37,17 @@ public class UserService {
         return userMapper.toModel(entity);
     }
 
-    public User update(User userWithId) {
-        userRepo.findById(userWithId.getId())
-                .orElseThrow(() -> new ResourceNotFoundException("User id: "+userWithId.getId()));
-        var entity = userRepo.save(userMapper.toEntity(userWithId));
+    public User update(User user) {
+        findUserEntityIfExists(user.getId());
+        var entity = userRepo.save(userMapper.toEntity(user));
         return userMapper.toModel(entity);
     }
 
     public void delete(Long userId) {
-        var userEntity = userRepo.findById(userId)
-                .orElseThrow(() -> new ResourceNotFoundException("User id: " + userId));
+        var userEntity = findUserEntityIfExists(userId);
+        var userHasCategories = !categoryService.findCategories(userId).isEmpty();
+        if (userHasCategories)
+            throw new TimeTrackerOperationException("User has Categories registered and cannot be deleted");
         userRepo.delete(userEntity);
     }
 
@@ -61,56 +57,14 @@ public class UserService {
     }
 
     public User findUserById(Long id) {
-        var entity = userRepo.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("User id: " + id));
+        var entity = findUserEntityIfExists(id);
         return userMapper.toModel(entity);
     }
 
-    public Category createCategory(Long userId, Category category) {
-        var userEntity = userRepo.findById(userId)
+
+    private UserEntity findUserEntityIfExists(Long userId) {
+        return userRepo.findById(userId)
                 .orElseThrow(() -> new ResourceNotFoundException("User id: " + userId));
-        var categoriesEntityOfThisUser = categoryRepo.retrieveByUser(userId);
-        var isNameAlreadyRegistered = categoriesEntityOfThisUser.stream()
-                .anyMatch(cat -> cat.getName().equals(category.getName()));
-        if (isNameAlreadyRegistered) throw new ResourceAlreadyRegisteredException("Category name already exists");
-        var categoryEntity = categoryMapper.toEntity(category);
-        categoryEntity.setUser(userEntity);
-        categoryEntity = categoryRepo.save(categoryEntity);
-        return categoryMapper.toModel(categoryEntity);
-    }
-
-    public Category update(Long userId, Category category) {
-        findCategoryEntityIfExists(userId, category.getId());
-        var categoryEntity = categoryMapper.toEntity(category);
-        categoryEntity = categoryRepo.save(categoryEntity);
-        return categoryMapper.toModel(categoryEntity);
-    }
-
-    public void delete(Long userId, Long categoryId) {
-        var categoryEntity = findCategoryEntityIfExists(userId, categoryId);
-        var isCategoryUsedInTimeEntries = !timeEntryRepo.retrieveByCategory(categoryId).isEmpty();
-        if (isCategoryUsedInTimeEntries)
-            throw new TimeTrackerOperationException("Category is associated to existing time entry and cannot be deleted");
-        categoryRepo.delete(categoryEntity);
-    }
-
-    public List<Category> findCategories(Long userId) {
-        return categoryRepo.retrieveByUser(userId).stream()
-                .map(categoryMapper::toModel).collect(toList());
-    }
-
-    public Category findCategoryById(Long userId, Long categoryId) {
-        return categoryMapper.toModel( findCategoryEntityIfExists(userId, categoryId) );
-    }
-
-
-    private CategoryEntity findCategoryEntityIfExists(Long userId, Long categoryId) {
-        userRepo.findById(userId)
-                .orElseThrow(() -> new ResourceNotFoundException("User id: " + userId));
-        return categoryRepo.retrieveByUser(userId).stream()
-                .filter(cat -> cat.getId().intValue() == categoryId.intValue())
-                .findFirst()
-                .orElseThrow(() -> new ResourceNotFoundException("Category id: " + categoryId));
     }
 
 }
